backend/script/requirements.txt
# UN Comtrade Data Fetching Pipeline Dependencies

requests>=2.28.0
pandas>=1.5.0
python-dotenv>=1.0.0
openpyxl>=3.0.0  # For reading Excel files
